- foundation for a better collision info management.
The way this is handled is the main reason why Build maps are size limited, but since it is extremely invasive it needs to be taken out of the game code piece by piece, this is the framework for doing this for Duke.

+++ source/games/duke/src/duke3d.h
@@ -88,7 +88,7 @@ struct Dispatcher
 	bool (*floorspace)(int sectnum);
 	void (*addweapon)(struct player_struct *p, int weapon);
 	void (*hitradius)(short i, int  r, int  hp1, int  hp2, int  hp3, int  hp4);
-	int  (*movesprite)(short spritenum, int xchange, int ychange, int zchange, unsigned int cliptype);
+	int  (*movesprite)(int spritenum, int xchange, int ychange, int zchange, unsigned int cliptype);
 	void (*lotsofmoney)(spritetype *s, short n);
 	void (*lotsofmail)(spritetype *s, short n);
 	void (*lotsofpaper)(spritetype *s, short n);
